// src/pinecone-generated-ts-fetch-alpha/db_control/models/IndexModel.ts
export interface IndexModel {
    /**
     * The name of the index. Resource name must be 1-45 characters long, start and end with an alphanumeric character, and consist only of lower case alphanumeric characters or '-'.
     * @type {string}
     * @memberof IndexModel
     */
    name: string;
    /**
     * The URL address where the index is hosted.
     * @type {string}
     * @memberof IndexModel
     */
    host: string;
    /**
     * The private endpoint URL of an index.
     * @type {string}
     * @memberof IndexModel
     */
    privateHost?: string;
    /**
     * 
     * @type {IndexModelStatus}
     * @memberof IndexModel
     */
    status: IndexModelStatus;
    /**
     * 
     * @type {IndexDeployment}
     * @memberof IndexModel
     */
    deployment: IndexDeployment;
    /**
     * 
     * @type {ReadCapacityResponse}
     * @memberof IndexModel
     */
    readCapacity?: ReadCapacityResponse;
    /**
     * The name of the collection this index was created from, if any.
     * @type {string}
     * @memberof IndexModel
     */
    sourceCollection?: string;
    /**
     * The ID of the backup this index was restored from, if any.
     * @type {string}
     * @memberof IndexModel
     */
    sourceBackupId?: string;
    /**
     * The ID of the customer-managed encryption key (CMEK) used to encrypt this index, if any.
     * @type {string}
     * @memberof IndexModel
     */
    cmekId?: string;
    /**
     * 
     * @type {IndexSchema}
     * @memberof IndexModel
     */
    schema: IndexSchema;
    /**
     * Custom user tags added to an index. Keys must be 80 characters or less. Values must be 120 characters or less. Keys must be alphanumeric, '_', or '-'.  Values must be alphanumeric, ';', '@', '_', '-', '.', '+', or ' '. To unset a key, set the value to be an empty string.
     * @type {{ [key: string]: string; }}
     * @memberof IndexModel
     */
    tags?: { [key: string]: string; };
    /**
     * Whether [deletion protection](http://docs.pinecone.io/guides/manage-data/manage-indexes#configure-deletion-protection) is enabled/disabled for the index.
     * Possible values: `disabled` or `enabled`.
     * @type {string}
     * @memberof IndexModel
     */
    deletionProtection: string;
}