// src/pinecone-generated-ts-fetch-alpha/db_control/models/BackupModel.ts
export interface BackupModel {
    /**
     * Unique identifier for the backup.
     * @type {string}
     * @memberof BackupModel
     */
    backupId: string;
    /**
     * Name of the index from which the backup was taken.
     * @type {string}
     * @memberof BackupModel
     */
    sourceIndexName: string;
    /**
     * ID of the index from which the backup was taken.
     * @type {string}
     * @memberof BackupModel
     */
    sourceIndexId: string;
    /**
     * Optional user-defined name for the backup.
     * @type {string}
     * @memberof BackupModel
     */
    name?: string;
    /**
     * Optional description providing context for the backup.
     * @type {string}
     * @memberof BackupModel
     */
    description?: string;
    /**
     * Current status of the backup.
     * @type {string}
     * @memberof BackupModel
     */
    status: string;
    /**
     * Cloud provider where the backup is stored.
     * @type {string}
     * @memberof BackupModel
     */
    cloud: string;
    /**
     * Cloud region where the backup is stored.
     * @type {string}
     * @memberof BackupModel
     */
    region: string;
    /**
     * 
     * @type {IndexSchema}
     * @memberof BackupModel
     */
    schema?: IndexSchema;
    /**
     * Total number of records in the backup.
     * @type {number}
     * @memberof BackupModel
     */
    recordCount?: number;
    /**
     * Number of namespaces in the backup.
     * @type {number}
     * @memberof BackupModel
     */
    namespaceCount?: number;
    /**
     * Size of the backup in bytes.
     * @type {number}
     * @memberof BackupModel
     */
    sizeBytes?: number;
    /**
     * Custom user tags added to an index. Keys must be 80 characters or less. Values must be 120 characters or less. Keys must be alphanumeric, '_', or '-'.  Values must be alphanumeric, ';', '@', '_', '-', '.', '+', or ' '. To unset a key, set the value to be an empty string.
     * @type {{ [key: string]: string; }}
     * @memberof BackupModel
     */
    tags?: { [key: string]: string; };
    /**
     * Timestamp when the backup was created.
     * @type {Date}
     * @memberof BackupModel
     */
    createdAt?: Date;
}